Fixed: multiple compiler warnings and an error for linux build

+++ 0.56.b/graysvr/CCharNPCPet.cpp
@@ -367,9 +367,9 @@ bool CChar::NPC_OnHearPetCmdTarg( int iCmd, CChar * pSrc, CObjBase * pObj, const
 			{
 				if ( IsSetEF(EF_PetSlots) && !pSrc->IsPriv(PRIV_GM) )
 				{
-					short int iFollowerSlotsNeeded = maximum(GetDefNum("FOLLOWERSLOTS", true),1);
-					short int iCurFollower = pCharTarg->GetDefNum("CURFOLLOWER", true);
-					short int iMaxFollower = pCharTarg->GetDefNum("MAXFOLLOWER", true);
+					short int iFollowerSlotsNeeded = static_cast<short>(maximum(GetDefNum("FOLLOWERSLOTS", true),1));
+					short int iCurFollower = static_cast<short>(pCharTarg->GetDefNum("CURFOLLOWER", true));
+					short int iMaxFollower = static_cast<short>(pCharTarg->GetDefNum("MAXFOLLOWER", true));
 
 					if ((iCurFollower + iFollowerSlotsNeeded) > iMaxFollower )
 					{
@@ -492,8 +492,8 @@ void CChar::NPC_PetClearOwners()
 		{
 			CChar * pPetOwner = pPetMemory->m_uidLink.CharFind();
 
-			short int iFollowerSlotsNeeded = maximum(GetDefNum("FOLLOWERSLOTS", true),1);
-			short int iCurFollower = pPetOwner->GetDefNum("CURFOLLOWER", true);
+			short int iFollowerSlotsNeeded = static_cast<short>(maximum(GetDefNum("FOLLOWERSLOTS", true),1));
+			short int iCurFollower = static_cast<short>(pPetOwner->GetDefNum("CURFOLLOWER", true));
 			short int iSetFollower = iCurFollower - iFollowerSlotsNeeded;
 			if ( iSetFollower < 0 )
 				iSetFollower = 0;
@@ -539,8 +539,8 @@ bool CChar::NPC_PetSetOwner( CChar * pChar )
 
 	if (IsSetEF(EF_PetSlots))
 	{
-		short int iFollowerSlotsNeeded = maximum(GetDefNum("FOLLOWERSLOTS", true),1);
-		short int iCurFollower = pChar->GetDefNum("CURFOLLOWER", true);
+		short int iFollowerSlotsNeeded = static_cast<short>(maximum(GetDefNum("FOLLOWERSLOTS", true),1));
+		short int iCurFollower = static_cast<short>(pChar->GetDefNum("CURFOLLOWER", true));
 		short int iSetFollower = iCurFollower + iFollowerSlotsNeeded;
 		if ( iSetFollower > 255 )
 			iSetFollower = 255;		// Max value that clients can show on char status
